Side quest to fix pointers. & => ^ now
