+++ prg/main.txt
@@ -1,5 +1,5 @@
 // External function
-extern "C" fn printf: s char&, ...;
+extern "C" fn printf: s char^, ...;
 
 //simple assignment
 let a = 3;
@@ -25,7 +25,7 @@ fn i32 foo:
 
 higher = higher + 1;
 
-fn main: argc i32, argv char& {
+fn main: argc i32, argv char^ {
     let x = 0;
     let y = 14;
